Moved DocumentServices to own folder.
